Refactor Client instantiation to support a factory/service location pattern for extensibility
Add in a well-known factory location to support instantiating an
instance of the EasyPost.Client class using a specific API key.

Provide support for the ClientFactory to take either a single instance of a
custom EasyPost.IClientFactory or a delegated factory method to defer the
provision of the factory object so as to be useful for DI/IoC implementations.

This will accommodate a broader range of use-cases and their respective
implementations by allowing more finely grained control of how the EasyPost.Client
class is instantiated.  One such example is the integration of EasyPost
into a multi-tenanted environment where individual tenants bring their own EasyPost
account (and API key).

The initialization of the EasyPost.ClientFactory should occur once at application
startup; further management of concurrency should be implemented in the
actual EasyPost.IClientFactory class.

Update documentation to reflect new configuration options.