Extend the Client object to accept ClientConfiguration options for apiKey assigment
This provides the ability to pass the API key and URL on a per-client basis,
making the library more flexible in a multi-account environment.

Additionally, refactored the base url value to an internally visible constant;
this provides consistent assignment of the value in the Client + ClientConfiguration classes
while maintaining backward compatibility.