Refactor Client/ClientManager to avoid fixation of single Client object across all callers/threads.
Removed unnecessary overloads of SetCurrent/SetDefault in ClientManager; also removed static ApiKey
property from Client as it creates issues with circular references back to ClientManager in some cases.

Updated ClientManager to maintain reference to factory delegate, not materialized Client object.
This allows a concurrent environment to utilize a new Client object on a per-call basis,
which in turn allows for multiple API Keys to be utilized within a single hosted environment.

Update ClientManager.Current property to be a method 'Build' to align with intent and implementation.
'Current' suggests a single instance; update usages.

Added tests to cover the ClientManager implementation; fix existing assertions
to be more properly aligned with the scenarios.