Fix issues with the port to v107. As mentioned here:
https://restsharp.dev/v107/#restclient-lifecycle

The RestClient needs to be a singleton instance now, not an instance per request so we introduce a RestClient cache to cache them per API endpoint.

Also since RestSharp v107 removed all the Sync functions from the API, it is incorrect to call an AsyncFunction and return the result using GetAwaiter().GetResult() as that WILL deadlock at some point. To do this correctly, you need to use a wrapper to run the functions Async and wait for the result. The wrapper in question is derived from the code in Rebus:

https://github.com/rebus-org/Rebus